refactor: web 프로젝트 구조 리팩터링
