+++ apps/web/src/components/molecules/DetailNavigation/index.module.scss
