+++ apps/web/src/components/molecules/DetailNavigation/index.tsx
