+++ apps/web/src/components/molecules/ImageCard/index.module.scss
