+++ apps/web/src/components/organisms/FAQSection/index.module.scss
