+++ apps/web/src/components/organisms/FAQSection/index.tsx
