+++ apps/web/src/components/organisms/GuidelineSection/index.module.scss
