+++ apps/web/src/components/organisms/JobCard/index.module.scss
