+++ apps/web/src/components/organisms/ProjectCards/index.module.scss
