+++ apps/web/src/components/organisms/ProjectsSlider/index.module.scss
