+++ apps/web/src/components/organisms/ProjectsSlider/index.tsx
