+++ apps/web/src/components/organisms/SponsorSection/index.module.scss
