+++ apps/web/src/components/organisms/SponsorSection/index.tsx
