+++ apps/web/src/components/pages/AboutPage/index.module.scss
@@ -46,13 +46,6 @@
   }
 }
 
-.counterCardWrapper {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 24px;
-  width: 100%;
-}
-
 .dndExperienceWrapper {
   display: flex;
   flex-direction: column;
