+++ packages/ui/vite.config.mts
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import { resolve } from 'path';
 
 import react from '@vitejs/plugin-react';
